requirements.txt
Django==1.11.9
channels==1.1.8
asgi-redis==1.4.3
Pillow==5.0.0
djangorestframework==3.7.7
coinbase==2.0.6
